Move fixtures to where they're used
It's ok to keep this in the same class they're used as long as they're not used in multiple places. This aligns it with other Laravel libraries.
