--- tests/Fixtures/CashierTestControllerStub.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Laravel\Cashier\Tests\Fixtures;
-
-use Laravel\Cashier\Http\Controllers\WebhookController;
-
-class CashierTestControllerStub extends WebhookController
-{
-    protected function eventExistsOnStripe($id)
-    {
-        return true;
-    }
-}
