--- tests/Fixtures/WebhookControllerTestStub.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Laravel\Cashier\Tests\Fixtures;
-
-use Laravel\Cashier\Http\Controllers\WebhookController;
-
-class WebhookControllerTestStub extends WebhookController
-{
-    public function handleChargeSucceeded()
-    {
-        $_SERVER['__received'] = true;
-    }
-
-    protected function eventExistsOnStripe($id)
-    {
-        return true;
-    }
-}
